# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TavernOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AbilityScores } from "./abilityScores";
import type { CharacterSheetData } from "./characterSheetData";

export interface Character {
  id: string;
  campaignId: string;
  userId: string;
  name: string;
  race?: string;
  class?: string;
  background?: string;
  level: number;
  hp: number;
  maxHp: number;
  tempHp?: number;
  ac: number;
  speed: number;
  initiativeBonus?: number;
  stats: AbilityScores;
  sheetData?: CharacterSheetData;
  tokenColor?: string;
  isNpc?: boolean;
  createdAt: Date;
}