# lib/api-zod/src/generated/types/createCharacterRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TavernOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AbilityScores } from "./abilityScores";
import type { CreateCharacterRequestSheetData } from "./createCharacterRequestSheetData";

export interface CreateCharacterRequest {
  name: string;
  race?: string;
  class?: string;
  background?: string;
  level?: number;
  hp?: number;
  maxHp?: number;
  ac?: number;
  speed?: number;
  stats: AbilityScores;
  sheetData?: CreateCharacterRequestSheetData;
  tokenColor?: string;
  isNpc?: boolean;
}